update: make index sitemap static
- update robots file to support compress sitemap static
